move blocking to new module
